Split buildDialogContent() to separate module
