+++ chrome-ext/src/manifest.json
@@ -32,6 +32,7 @@
         "js/tingle.js",
         "js/mustache.js",
         "js/hots-dialog.js",
+        "js/hots-dialog-builder.js",
         "js/hots-dialog-renderers.js",
         "js/hots-dialog-util.js"
       ]
